Struct burn_autodiff::grads::Gradients

pub struct Gradients { /* private fields */ }

Gradients container used during the backward pass.

Implementations

impl Gradients

pub fn new<B: Backend, const D: usize>( root_node: Arc<Node>, root_tensor: <B as Backend>::TensorPrimitive<D> ) -> Self

Creates a new gradients container.

pub fn consume<B: Backend, const D: usize>( &mut self, node: &Arc<Node> ) -> <B as Backend>::TensorPrimitive<D>

Consumes the gradients for a given tensor.

Each tensor should be consumed exactly 1 time if its gradients are only required during the backward pass, otherwise, it may be consume multiple times.

pub fn remove<B: Backend, const D: usize>( &mut self, tensor: &AutodiffTensor<B, D> ) -> Option<<B as Backend>::TensorPrimitive<D>>

Removes a grad tensor from the container.

pub fn get<B: Backend, const D: usize>( &self, tensor: &AutodiffTensor<B, D> ) -> Option<<B as Backend>::TensorPrimitive<D>>

Gets a grad tensor from the container.

pub fn register<B: Backend, const D: usize>( &mut self, node: Arc<Node>, value: <B as Backend>::TensorPrimitive<D> )

Register a grad tensor in the container.

If the tensor already exists, add both tensors together before saving the result.